// client/src/redux/home/constants.ts
export enum constants {
    GET_FLASH_DEALS_START = 'GET_FLASH_DEALS_START',
    GET_FLASH_DEALS_SUCCESS = 'GET_FLASH_DEALS_SUCCESS',
    GET_NEW_ARRIVALS_START = "GET_NEW_ARRIVALS_START",
    GET_NEW_ARRIVALS_SUCCESS = "GET_NEW_ARRIVALS_SUCCESS",
    GET_TOP_RATING_START = "GET_TOP_RATING_START",
    GET_TOP_RATING_SUCCESS = "GET_TOP_RATING_SUCCESS",
    GET_TOP_CATEGORIES_START = "GET_TOP_CATEGORIES_START",
    GET_TOP_CATEGORIES_SUCCESS = "GET_TOP_CATEGORIES_SUCCESS",
    RESET_HOME_STATE = 'RESET_HOME_STATE'
}